# contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vis Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/slides": {
      "get": {
        "tags": [
          "slides"
        ],
        "summary": "List Decks",
        "operationId": "list_decks_api_v1_projects__project_id__slides_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "slides"
        ],
        "summary": "Create Deck",
        "operationId": "create_deck_api_v1_projects__project_id__slides_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSlideDeck"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/slides/{deck_id}/content": {
      "get": {
        "tags": [
          "slides"
        ],
        "summary": "Export Deck",
        "operationId": "export_deck_api_v1_projects__project_id__slides__deck_id__content_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Deck Id"
            }
          },
          {
            "name": "freeze",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Freeze"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlideDeckContent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ControlsMode": {
        "type": "string",
        "enum": [
          "language",
          "panel",
          "hybrid"
        ],
        "title": "ControlsMode"
      },
...
      "CreateSlideDeck": {
        "properties": {
          "request_id": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Request Id"
          },
          "content": {
            "$ref": "#/components/schemas/SlideDeckContent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "content"
        ],
        "title": "CreateSlideDeck"
      },
      "Dataset": {
        "properties": {
          "contains_demo_data": {
            "type": "boolean",
            "title": "Contains Demo Data",
            "default": false
          },
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "source_kind": {
            "type": "string",
            "enum": [
              "upload",
              "analysis_platform"
            ],
            "title": "Source Kind"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "source_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision"
          },
          "revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "uploading",
              "processing",
              "ready",
              "failed"
            ],
            "title": "State",
            "default": "uploading"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/ObjectDescription"
            },
            "type": "array",
            "title": "Objects"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/ObjectRelationship"
            },
            "type": "array",
            "title": "Relationships"
          },
          "interpretation_status": {
            "type": "string",
            "enum": [
              "pending",
              "source_provided",
              "interpreted",
              "unavailable"
            ],
            "title": "Interpretation Status",
            "default": "pending"
          },
          "notices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notices"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_id",
          "project_id",
          "name",
          "source_kind",
          "source_id",
          "created_at",
          "updated_at"
        ],
        "title": "Dataset"
      },
...
      "FigureSize": {
        "properties": {
          "width": {
            "type": "number",
            "multipleOf": 0.01,
            "maximum": 30.0,
            "minimum": 2.0,
            "title": "Width"
          },
          "height": {
            "type": "number",
            "multipleOf": 0.01,
            "maximum": 30.0,
            "minimum": 2.0,
            "title": "Height"
          },
          "unit": {
            "type": "string",
            "const": "in",
            "title": "Unit",
            "default": "in"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "width",
          "height"
        ],
        "title": "FigureSize"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlotResultSummary": {
        "properties": {
          "reference_images": {
            "items": {
              "$ref": "#/components/schemas/ReferenceImage"
            },
            "type": "array",
            "title": "Reference Images"
          },
          "figure_size": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FigureSize"
              },
              {
                "type": "null"
              }
            ]
          },
          "contains_demo_data": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contains Demo Data"
          },
          "plot_id": {
            "type": "string",
            "title": "Plot Id"
          },
          "version_id": {
            "type": "string",
            "title": "Version Id"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "demo",
              "r"
            ],
            "title": "Execution Mode"
          },
          "preview": {
            "$ref": "#/components/schemas/ArtifactReference"
          },
          "controls_mode": {
            "$ref": "#/components/schemas/ControlsMode"
          },
          "controls": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/NumberControl"
                },
                {
                  "$ref": "#/components/schemas/ChoiceControl"
                },
                {
                  "$ref": "#/components/schemas/TextControl"
                },
                {
                  "$ref": "#/components/schemas/BooleanControl"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "boolean": "#/components/schemas/BooleanControl",
                  "choice": "#/components/schemas/ChoiceControl",
                  "number": "#/components/schemas/NumberControl",
                  "text": "#/components/schemas/TextControl"
                }
              }
            },
            "type": "array",
            "title": "Controls"
          },
          "control_groups": {
            "items": {
              "$ref": "#/components/schemas/ControlGroup"
            },
            "type": "array",
            "title": "Control Groups"
          },
          "parameter_schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Parameter Schema Version",
            "default": "1.0"
          },
          "parameter_updates_available": {
            "type": "boolean",
            "title": "Parameter Updates Available",
            "default": false
          },
          "data_summary": {
            "type": "string",
            "title": "Data Summary",
            "default": "Data provenance was not recorded for this version."
          },
          "data_used": {
            "items": {
              "$ref": "#/components/schemas/PlotDataObject"
            },
            "type": "array",
            "title": "Data Used"
          },
          "input_objects": {
            "items": {
              "$ref": "#/components/schemas/ObjectReference"
            },
            "type": "array",
            "title": "Input Objects"
          },
          "analysis_results": {
            "items": {
              "$ref": "#/components/schemas/AnalysisResult"
            },
            "type": "array",
            "title": "Analysis Results"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationSummary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plot_id",
          "version_id",
          "execution_mode",
          "preview",
          "controls_mode",
          "validation"
        ],
        "title": "PlotResultSummary"
      },
...
      "PresentationSettings": {
        "properties": {
          "aspect_ratio": {
            "type": "string",
            "const": "16:9",
            "title": "Aspect Ratio",
            "default": "16:9"
          },
          "theme": {
            "type": "string",
            "enum": [
              "paper",
              "midnight"
            ],
            "title": "Theme",
            "default": "paper"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PresentationSettings"
      },
...
      "ReferenceImage": {
        "properties": {
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "media_type": {
            "type": "string",
            "const": "image/png",
            "title": "Media Type",
            "default": "image/png"
          },
          "width": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Height"
          },
          "byte_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Byte Size"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "links": {
            "$ref": "#/components/schemas/ReferenceImageLinks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "image_id",
          "project_id",
          "name",
          "width",
          "height",
          "byte_size",
          "created_at",
          "links"
        ],
        "title": "ReferenceImage"
      },
...
      "ReportContent": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version",
            "default": "2.0"
          },
          "kind": {
            "type": "string",
            "enum": [
              "report",
              "slides"
            ],
            "title": "Kind",
            "default": "report"
          },
          "presentation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresentationSettings"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/ReportDataset"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Datasets"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ReportSection"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Sections",
            "description": "Document order: a level-1 section is followed by its ordered level-2 subsections."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ReportContent"
      },
      "ReportDataset": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Dataset Id"
          },
          "revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_id"
        ],
        "title": "ReportDataset"
      },
      "ReportDocument": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "content": {
            "$ref": "#/components/schemas/ReportContent"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/Dataset"
            },
            "type": "array",
            "title": "Datasets"
          },
          "figure_bindings": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Figure Bindings"
          },
          "figures": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PlotResultSummary"
            },
            "type": "object",
            "title": "Figures"
          },
          "images": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ReferenceImage"
            },
            "type": "object",
            "title": "Images"
          },
          "edits": {
            "items": {
              "$ref": "#/components/schemas/ReportEdit"
            },
            "type": "array",
            "title": "Edits"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ReportMessage"
            },
            "type": "array",
            "title": "Messages"
          },
          "stale_text_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Stale Text Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "report_id",
          "project_id",
          "title",
          "revision",
          "updated_at",
          "created_at",
          "content",
          "datasets",
          "figures",
          "images",
          "edits"
        ],
        "title": "ReportDocument"
      },
      "ReportEdit": {
        "properties": {
          "edit_id": {
            "type": "string",
            "title": "Edit Id"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "block_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Id"
          },
          "output_block_id": {
            "type": "string",
            "title": "Output Block Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "figure",
              "text",
              "discussion"
            ],
            "title": "Kind"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "awaiting_input",
              "awaiting_approval",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "response_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Text"
          },
          "dismissed": {
            "type": "boolean",
            "title": "Dismissed",
            "default": false
          },
          "message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "assistant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssistantTurnAccepted"
              },
              {
                "type": "null"
              }
            ]
          },
          "assistant_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssistantTurnSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlotRunAccepted"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlotRunSnapshot"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "edit_id",
          "section_id",
          "output_block_id",
          "kind",
          "prompt",
          "status",
          "created_at"
        ],
        "title": "ReportEdit"
      },
...
      "ReportList": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "reports": {
            "items": {
              "$ref": "#/components/schemas/ReportSummary"
            },
            "type": "array",
            "title": "Reports"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reports",
          "total",
          "offset"
        ],
        "title": "ReportList"
      },
      "ReportMessage": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportSelection"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "awaiting_input",
              "awaiting_approval",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "phase": {
            "type": "string",
            "enum": [
              "planning",
              "applying",
              "generating",
              "finished"
            ],
            "title": "Phase",
            "default": "planning"
          },
          "response_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Text"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlannerQuestions"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_edit_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Edit Id"
          },
          "completed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Completed Actions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message_id",
          "prompt",
          "status",
          "created_at"
        ],
        "title": "ReportMessage"
      },
...
      "ReportSection": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "level": {
            "type": "integer",
            "enum": [
              1,
              2
            ],
            "title": "Level",
            "default": 1
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "slide": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SlideSettings"
              },
              {
                "type": "null"
              }
            ]
          },
          "blocks": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ReportTextBlock"
                },
                {
                  "$ref": "#/components/schemas/ReportFigureBlock"
                },
                {
                  "$ref": "#/components/schemas/ReportTableBlock"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "figure": "#/components/schemas/ReportFigureBlock",
                  "table": "#/components/schemas/ReportTableBlock",
                  "text": "#/components/schemas/ReportTextBlock"
                }
              }
            },
            "type": "array",
            "maxItems": 200,
            "title": "Blocks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "ReportSection"
      },
...
      "ReportSummary": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "report_id",
          "project_id",
          "title",
          "revision",
          "updated_at"
        ],
        "title": "ReportSummary"
      },
...
      "Slide": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "elements": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ReportTextBlock"
                },
                {
                  "$ref": "#/components/schemas/ReportFigureBlock"
                },
                {
                  "$ref": "#/components/schemas/ReportTableBlock"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "figure": "#/components/schemas/ReportFigureBlock",
                  "table": "#/components/schemas/ReportTableBlock",
                  "text": "#/components/schemas/ReportTextBlock"
                }
              }
            },
            "type": "array",
            "maxItems": 200,
            "title": "Elements"
          },
          "settings": {
            "$ref": "#/components/schemas/SlideSettings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "Slide"
      },
      "SlideDeckContent": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/ReportDataset"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Datasets"
          },
          "presentation": {
            "$ref": "#/components/schemas/PresentationSettings"
          },
          "slides": {
            "items": {
              "$ref": "#/components/schemas/Slide"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Slides"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "SlideDeckContent"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}